// Assets-Uncapped/front-portal-develop/src/services/api/agreements/models/RepaymentLinkedTransaction.ts
/* tslint:disable */
/* eslint-disable */
/**
 * api-agreements
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: Unknown version
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RepaymentLinkedTransaction
 */
export interface RepaymentLinkedTransaction {
    /**
     * 
     * @type {Date}
     * @memberof RepaymentLinkedTransaction
     */
    date?: Date;
    /**
     * 
     * @type {number}
     * @memberof RepaymentLinkedTransaction
     */
    amount?: number;
    /**
     * 
     * @type {string}
     * @memberof RepaymentLinkedTransaction
     */
    status?: RepaymentLinkedTransactionStatusEnum;
}


/**
 * @export
 */
export const RepaymentLinkedTransactionStatusEnum = {
    Requested: 'REQUESTED',
    Pending: 'PENDING',
    Paid: 'PAID',
    Failed: 'FAILED',
    Cancelled: 'CANCELLED',
    Unknown: 'UNKNOWN'
} as const;
export type RepaymentLinkedTransactionStatusEnum = typeof RepaymentLinkedTransactionStatusEnum[keyof typeof RepaymentLinkedTransactionStatusEnum];


/**
 * Check if a given object implements the RepaymentLinkedTransaction interface.
 */
export function instanceOfRepaymentLinkedTransaction(value: object): value is RepaymentLinkedTransaction {
    return true;
}

export function RepaymentLinkedTransactionFromJSON(json: any): RepaymentLinkedTransaction {
    return RepaymentLinkedTransactionFromJSONTyped(json, false);
}

export function RepaymentLinkedTransactionFromJSONTyped(json: any, ignoreDiscriminator: boolean): RepaymentLinkedTransaction {
    if (json == null) {
        return json;
    }
    return {
        
        'date': json['date'] == null ? undefined : (new Date(json['date'])),
        'amount': json['amount'] == null ? undefined : json['amount'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function RepaymentLinkedTransactionToJSON(json: any): RepaymentLinkedTransaction {
    return RepaymentLinkedTransactionToJSONTyped(json, false);
}

export function RepaymentLinkedTransactionToJSONTyped(value?: RepaymentLinkedTransaction | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'date': value['date'] == null ? undefined : ((value['date']).toISOString().substring(0,10)),
        'amount': value['amount'],
        'status': value['status'],
    };
}